Fixed upload

+++ app/models/upload.rb
@@ -27,7 +27,8 @@ class Upload
     User.all.each{|u|
       u.destroy if not u.login=="admin"
     }
-    models.each{|model|
+    models.each{|model|      
+      next unless File.exists?(File.join(Merb.root, "uploads", @directory, model.to_s.snake_case.pluralize))
       model.all.destroy!
       log.info("Destroying old records for #{model.to_s.plural} (if any)") if log
       log.info("Creating #{model.to_s.plural}") if log
